requirements.txt
roboflow
ultralytics==8.0.196
pillow